requirements.txt
discord.py>=2.5.0
aiohttp>=3.7.4